requirements.txt
tensorflow==2.20.0
gradio==4.44.0
pillow==10.1.0
numpy==1.26.4
